+++ datafusion/expr/Cargo.toml
@@ -36,6 +36,8 @@ name = "datafusion_expr"
 path = "src/lib.rs"
 
 [features]
+default = ["franz"]
+franz = []
 
 [dependencies]
 ahash = { workspace = true }
