+++ datafusion/expr/src/logical_plan/mod.rs
@@ -24,6 +24,9 @@ mod plan;
 mod statement;
 pub mod tree_node;
 
+#[cfg(feature = "franz")]
+mod franz;
+
 pub use builder::{
     build_join_schema, table_scan, union, wrap_projection_for_join_if_necessary,
     LogicalPlanBuilder, UNNAMED_TABLE,
