+++ datafusion/physical-plan/src/windows/mod.rs
@@ -46,15 +46,18 @@ use datafusion_physical_expr::{
 use itertools::Itertools;
 
 mod bounded_window_agg_exec;
+
+#[cfg(feature = "franz")]
 pub mod stream_window_agg_exec;
+#[cfg(feature = "franz")]
+pub use stream_window_agg_exec::{FranzWindowType, FranzWindowExec};
+
 mod window_agg_exec;
 
 pub use bounded_window_agg_exec::BoundedWindowAggExec;
 pub use datafusion_physical_expr::window::{
     BuiltInWindowExpr, PlainAggregateWindowExpr, WindowExpr,
 };
-pub use stream_window_agg_exec::FranzWindowExec;
-pub use stream_window_agg_exec::FranzWindowType;
 pub use window_agg_exec::WindowAggExec;
 
 /// Build field from window function and add it into schema
